8 byte correction in malloc in test files

+++ tests/test_nsg_search.cpp
@@ -18,7 +18,7 @@ void load_data(char* filename, float*& data, unsigned& num,
   std::ios::pos_type ss = in.tellg();
   size_t fsize = (size_t)ss;
   num = (unsigned)(fsize / (dim + 1) / 4);
-  data = new float[num * dim * sizeof(float)];
+  data = new float[(size_t) num * (size_t) dim * sizeof(float)];
 
   in.seekg(0, std::ios::beg);
   for (size_t i = 0; i < num; i++) {
